Investigate data stored in an instance of a class by executing an algorithm to solve an inverse problem

investigate method applied to SpecificModels

investigate method applied to SpecificMixed

investigate method applied to SpecificPredictors

investigate method applied to SpecificFixed

# S4 method for SpecificModels
investigate(x)

# S4 method for SpecificMixed
investigate(x, backend_1L_chr = "cmdstanr", new_dir_nm_1L_chr = "F_TS_Mdls")

# S4 method for SpecificPredictors
investigate(x)

# S4 method for SpecificFixed
investigate(x)

Arguments

x

An object of class SpecificFixed

backend_1L_chr

Backend (a character vector of length one), Default: 'cmdstanr'

new_dir_nm_1L_chr

New directory name (a character vector of length one), Default: 'F_TS_Mdls'

Value

x (An object of class SpecificMixed) X (Modelling project dataset, input parameters and complete fixed models results.) X (Modelling project dataset, input parameters and complete mixed models results.)