Candidate models lookup table

Create a new valid instance of the Candidate models lookup table

specific_models(x = make_pt_specific_models())

make_pt_specific_models(
  short_name_chr = character(0),
  long_name_chr = character(0),
  control_chr = character(0),
  family_chr = character(0),
  fn_chr = character(0),
  start_chr = character(0),
  predn_type_chr = character(0),
  tfmn_chr = character(0),
  tfmn_for_bnml_lgl = logical(0),
  fixed_acronym_chr = character(0),
  mixed_acronym_chr = character(0),
  mixed_type_chr = character(0),
  with_chr = character(0)
)

is_specific_models(x)

Arguments

x

An object of any type

short_name_chr

Short name (a character vector), Default: character(0)

long_name_chr

Long name (a character vector), Default: character(0)

control_chr

Control (a character vector), Default: character(0)

family_chr

Family (a character vector), Default: character(0)

fn_chr

Function (a character vector), Default: character(0)

start_chr

Start (a character vector), Default: character(0)

predn_type_chr

Prediction type (a character vector), Default: character(0)

tfmn_chr

Transformation (a character vector), Default: character(0)

tfmn_for_bnml_lgl

Transformation for binomial (a logical vector), Default: logical(0)

fixed_acronym_chr

Fixed acronym (a character vector), Default: character(0)

mixed_acronym_chr

Mixed acronym (a character vector), Default: character(0)

mixed_type_chr

Mixed type (a character vector), Default: character(0)

with_chr

With (a character vector), Default: character(0)

Value

A validated instance of the Candidate models lookup table A prototype for Candidate models lookup table A logical value, TRUE if a valid instance of the Candidate models lookup table

Details

Candidate models lookup table