Write report with record

write_rprt_with_rcrd() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write report with record. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour

write_rprt_with_rcrd(
  path_to_outp_fl_1L_chr,
  paths_ls,
  header_yaml_args_ls = NULL,
  rprt_lup = NULL,
  use_fake_data_1L_lgl = F,
  rprt_nm_1L_chr = "AAA_TTU_MDL_CTG",
  rcrd_nm_1L_chr = "AAA_RPRT_WRTNG_MTH",
  reference_1L_int = NULL,
  start_at_int = c(2, 1),
  output_type_1L_chr = "PDF",
  rprt_output_type_1L_chr = "PDF",
  nbr_of_digits_1L_int = 2L,
  abstract_args_ls = NULL,
  main_rprt_append_ls = NULL,
  rcrd_rprt_append_ls = NULL
)

Arguments

path_to_outp_fl_1L_chr

Path to output file (a character vector of length one)

paths_ls

Paths (a list)

header_yaml_args_ls

Header yaml arguments (a list), Default: NULL

rprt_lup

Report (a lookup table), Default: NULL

use_fake_data_1L_lgl

Use fake data (a logical vector of length one), Default: F

rprt_nm_1L_chr

Report name (a character vector of length one), Default: 'AAA_TTU_MDL_CTG'

rcrd_nm_1L_chr

Record name (a character vector of length one), Default: 'AAA_RPRT_WRTNG_MTH'

reference_1L_int

Reference (an integer vector of length one), Default: NULL

start_at_int

Start at (an integer vector), Default: c(2, 1)

output_type_1L_chr

Output type (a character vector of length one), Default: 'PDF'

rprt_output_type_1L_chr

Report output type (a character vector of length one), Default: 'PDF'

nbr_of_digits_1L_int

Number of digits (an integer vector of length one), Default: 2

abstract_args_ls

Abstract arguments (a list), Default: NULL

main_rprt_append_ls

Main report append (a list), Default: NULL

rcrd_rprt_append_ls

Record report append (a list), Default: NULL