Write secondary analysis

write_scndry_analysis() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write secondary analysis. The function returns Output summary (a list).

Usage

write_scndry_analysis(
  valid_params_ls_ls,
  candidate_covar_nms_chr,
  path_params_ls,
  reference_1L_int,
  backend_1L_chr = "cmdstanr",
  candidate_predrs_chr = NULL,
  combinations_1L_lgl = F,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  cores_1L_int = 1L,
  depnt_var_max_val_1L_dbl = 0.99999,
  depnt_var_min_val_1L_dbl = 1e-05,
  existing_predrs_ls = NULL,
  max_nbr_of_covars_1L_int = integer(0),
  new_dir_nm_1L_chr = "F_TS_Mdls",
  options_chr = c("Y", "N"),
  predictors_lup = NULL,
  prefd_covars_chr = NA_character_,
  signft_covars_cdn_1L_chr = "any"
)

Arguments

valid_params_ls_ls

Valid parameters (a list of lists)

candidate_covar_nms_chr

Candidate covariate names (a character vector)

path_params_ls

Path parameters (a list)

reference_1L_int

Reference (an integer vector of length one)

backend_1L_chr

Backend (a character vector of length one), Default: 'cmdstanr'

candidate_predrs_chr

Candidate predictors (a character vector), Default: NULL

combinations_1L_lgl

Combinations (a logical vector of length one), Default: F

consent_1L_chr

Consent (a character vector of length one), Default: ''

consent_indcs_int

Consent indices (an integer vector), Default: 1

cores_1L_int

Cores (an integer vector of length one), Default: 1

depnt_var_max_val_1L_dbl

Dependent variable maximum value (a double vector of length one), Default: 0.99999

depnt_var_min_val_1L_dbl

Dependent variable minimum value (a double vector of length one), Default: 1e-05

existing_predrs_ls

Existing predictors (a list), Default: NULL

max_nbr_of_covars_1L_int

Maximum number of covariates (an integer vector of length one), Default: integer(0)

new_dir_nm_1L_chr

New directory name (a character vector of length one), Default: 'F_TS_Mdls'

options_chr

Options (a character vector), Default: c("Y", "N")

predictors_lup

Predictors (a lookup table), Default: NULL

prefd_covars_chr

Preferred covariates (a character vector), Default: 'NA'

signft_covars_cdn_1L_chr

Significant covariates condition (a character vector of length one), Default: 'any'

Value

Output summary (a list)